Module Morbig.ExtMenhirLib

val current_items : 'a Morbig__.Parser.MenhirInterpreter.env -> Morbig__.Parser.MenhirInterpreter.item list

type 'a status =

| AcceptedNow of 'a

| Fine

| Wrong

val close : 'a Morbig__.Parser.MenhirInterpreter.checkpoint -> 'a status

val accepted_token : 'a Morbig__.Parser.MenhirInterpreter.checkpoint -> (Morbig__.Parser.MenhirInterpreter.token * MenhirLib.IncrementalEngine.position * MenhirLib.IncrementalEngine.position) -> 'a status

val is_accepted_token : 'a Morbig__.Parser.MenhirInterpreter.checkpoint -> (Morbig__.Parser.MenhirInterpreter.token * MenhirLib.IncrementalEngine.position * MenhirLib.IncrementalEngine.position) -> bool

val accepted_raw_token : 'a Morbig__.Parser.MenhirInterpreter.checkpoint -> Morbig__.Parser.MenhirInterpreter.token -> 'a status

val finished : 'a Morbig__.Parser.MenhirInterpreter.checkpoint -> bool

finished checkpoint is true if the current checkpoint can move the LR(1) automaton to an accepting state with no extra input.

type nonterminal =

| AnyN : 'a Parser.MenhirInterpreter.nonterminal -> nonterminal

nonterminal_production p returns the non terminal of p. The nonterminals of Menhir API are a too precisely typed for our needs. Hence, we introduce an extential type for weaken this precision.

val nonterminal_of_production : Morbig__.Parser.MenhirInterpreter.production -> nonterminal

exception EmptyStack

type 'b top_symbol_processor = {

perform : a. ('a Morbig__.Parser.MenhirInterpreter.symbol * 'a) -> 'b;

}

val on_top_symbol : 'a Morbig__.Parser.MenhirInterpreter.env -> 'b top_symbol_processor -> 'b