Module Morbig__.JsonHelpers

val save_as_json : bool -> Stdlib.out_channel -> Morbig.CST.program -> unit

save_as_json simplified oc cst writes the concrete syntax tree cst to the out_channel oc. If simplified is true then location information is omitted, otherwise it is included in the json output.

val load_from_json : Stdlib.in_channel -> Morbig.CST.program

load_from_json ic retrieves a concrete syntax tree cst from the in_channel ic, assuming it has been generated using save_as_json false.

val save_as_dot : Stdlib.out_channel -> Morbig.CST.program -> unit

save_as_dot oc cst writes the concrete syntax tree cst to the out_channel oc using the DOT format.