Module Morbig__.QuoteRemoval

val on_string : string -> string

on_string s yields a copy of string s, with all quotes removed as described in the specification.

type backslash_automaton_state =

| Default

| Backslash

val backslash_as_in_doublequotes : string -> string

val remove_tabs_at_linestart : string -> string