Module Morbig__ExtPervasives

val nat_exp : int -> int -> int

val comment : (unit -> 'a) -> ('a -> unit) -> 'a

val string_of_channel : Stdlib.in_channel -> string

val split_list : ('a -> bool) -> 'a list -> 'a list list

val uniq : 'a list -> 'a list

val histogram : ('a -> 'b) -> 'a list -> ('a * int) list

val option_iter : 'a option -> ('a -> unit) -> unit

val option_map : 'a option -> ('a -> 'b) -> 'b option

val string_cut_at : int -> string -> string

exception InvalidSuffix of string * string

val string_split : int -> string -> string * string

val string_remove_suffix : string -> string -> string

val string_last_char : string -> char

val string_minus_last_char : string -> string

val preceded_by : int -> 'a -> 'a list -> bool

val preceding : 'a -> 'a list -> int

val string_to_char_list : string -> char list

string_to_char_list s turns a string s into a list of char.

val string_of_char_list : char list -> string

val count_end_character : char -> string -> int

val string_strip : string -> string

strip s returns a copy of s, without any final newline

val reduce : 'a -> ('a -> 'a -> 'a) -> 'a list -> 'a

val repeat : int -> (int -> 'a) -> 'a list

val take : int -> 'a list -> 'a list * 'a list

val take_until : ('a -> bool) -> 'a list -> 'a list * 'a list

val hashtbl_to_list : ('a, 'b) Stdlib.Hashtbl.t -> ('a * 'b) list

val list_nth_opt : 'a list -> int -> 'a option

val list_hd_opt : 'a list -> 'a option

module FirstSuccessMonad : sig ... end

val pp_string : Stdlib.Format.formatter -> string -> unit

val pp_to_print : (Stdlib.Format.formatter -> 'a) -> 'a

pp_to_print pp is the pretty-printer pp that, instead of taking any formatter, uses the std_formatter.

val pp_to_string : (Stdlib.Format.formatter -> 'a -> unit) -> 'a -> string

val lexing_make : string -> string -> Stdlib.Lexing.lexbuf

val (<$>) : 'a -> (unit -> unit) -> 'a

val list_last : 'a list -> 'a option

val newline_regexp : Str.regexp

val lines : string -> string list

val string_last_line : string -> string option