Module Morbig__HereDocument

HereDocument.Lexer handles here documents. It does:

internal registration of redirection operators (<< or <<-), and of the delimiter word that follows such an operator.

scannig of the contents of a series of here documents.

module Lexer : functor (U : sig ... end) -> sig ... end